requirements.txt
flask==2.1.3
gunicorn==20.0.4
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0
torchtext==0.12.0
speechbrain==0.5.12
pyannote.audio==2.0.1
pydub==0.25.1